Move vehicles into seperate folder
