+++ src/OpenLoco/Audio/Audio.cpp
@@ -9,9 +9,9 @@
 #include "../Objects/ObjectManager.h"
 #include "../Objects/SoundObject.h"
 #include "../Things/ThingManager.h"
-#include "../Things/Vehicle.h"
 #include "../Ui/WindowManager.h"
 #include "../Utility/Stream.hpp"
+#include "../Vehicles/Vehicle.h"
 #include "Channel.h"
 #include "MusicChannel.h"
 #include "VehicleChannel.h"
