+++ src/OpenLoco/Paint/PaintEntity.cpp
@@ -4,6 +4,7 @@
 #include "../Map/Tile.h"
 #include "../Things/Misc.h"
 #include "../Things/ThingManager.h"
+#include "../Vehicles/Vehicle.h"
 #include "Paint.h"
 #include "PaintVehicle.h"
