+++ src/OpenLoco/Vehicles/CreateVehicle.cpp
@@ -11,8 +11,8 @@
 #include "../Objects/SoundObject.h"
 #include "../Objects/TrackObject.h"
 #include "../Objects/VehicleObject.h"
+#include "../Things/ThingManager.h"
 #include "../Ui/WindowManager.h"
-#include "ThingManager.h"
 #include "Vehicle.h"
 #include <numeric>
 #include <utility>
