+++ src/OpenLoco/Windows/BuildVehicle.cpp
@@ -17,6 +17,7 @@
 #include "../Things/ThingManager.h"
 #include "../Ui/ScrollView.h"
 #include "../Ui/WindowManager.h"
+#include "../Vehicles/Vehicle.h"
 #include "../Widget.h"
 
 using namespace OpenLoco::Interop;
