+++ src/OpenLoco/Windows/ToolbarTop.cpp
@@ -16,10 +16,10 @@
 #include "../Objects/WaterObject.h"
 #include "../StationManager.h"
 #include "../Things/ThingManager.h"
-#include "../Things/Vehicle.h"
 #include "../TownManager.h"
 #include "../Ui/Dropdown.h"
 #include "../Ui/WindowManager.h"
+#include "../Vehicles/Vehicle.h"
 #include "ToolbarTopCommon.h"
 #include <map>
