+++ src/OpenLoco/Windows/ToolbarTopAlt.cpp
@@ -17,10 +17,10 @@
 #include "../S5/S5.h"
 #include "../StationManager.h"
 #include "../Things/ThingManager.h"
-#include "../Things/Vehicle.h"
 #include "../TownManager.h"
 #include "../Ui/Dropdown.h"
 #include "../Ui/WindowManager.h"
+#include "../Vehicles/Vehicle.h"
 #include "ToolbarTopCommon.h"
 
 using namespace OpenLoco::Interop;
